Force Aead::AeadImpl to be Send + Sync
When trying to use AeadCtxR/AeadCtxS in a generic context which requires
Send or Sync, downstream users need to add a bound to Aead::AeadImpl
even though this is a doc(hidden) item and thus an implementation detail
that may change afterwards.

All implementations of Aead provide an AeadImpl type that is already
Send + Sync so it doesn't hurt to add a trait bound there.